openfpga-physical/requirements.txt
pandas
numpy
envyaml
matplotlib
svgwrite
spydrnet
git+https://github.com/ganeshgore/spydrnet-physical.git